Remove unused dependencies and redundant getters, improve comments, and fix minor issues
- Removed unused dependencies from `pom.xml` files to clean up project configurations.
- Eliminated redundant getter methods from `WFLOP` for better code maintainability.
- Improved code comments for clarity in `PowerOutputCalculator`.
- Adjusted type casting in `WakeCalculatorJensen` for correctness.
- Fixed minor formatting issue in `README.md`.

+++ wflop-core/src/main/java/org/zafer/wflopcore/calculator/WakeCalculatorJensen.java
@@ -99,9 +99,9 @@ public class WakeCalculatorJensen {
 
     private double[] computeRotatedDistance(int from, int to, int angle) {
         double x1 = (from % dimension + 0.5) * gridWidth;
-        double y1 = (from / dimension + 0.5) * gridWidth;
+        double y1 = ((double) from / dimension + 0.5) * gridWidth;
         double x2 = (to % dimension + 0.5) * gridWidth;
-        double y2 = (to / dimension + 0.5) * gridWidth;
+        double y2 = ((double) to / dimension + 0.5) * gridWidth;
 
         double dx = x2 - x1;
         double dy = y2 - y1;
